Make CTAssetsPickerDemo use the CTAssetsPickerController pod.
Added a Podfile for the demo project with a reference to the current
directory. This will use the podspec in the current directory to build
the CTAssetsPickerController bundle.
Removed all CTAssetsPickerController/* files from the demo target.
These files will be compiled into the pod library that will be linked
to the demo by cocoapods.
Updated project settings to the ones recommended by Xcode.
Added the Pods directory to gitignore.